+++ src/core/EventPayload.ts
@@ -1,7 +1,11 @@
+import { v4 as uuidv4 } from 'uuid';
+
 /**
  * Event payload with data.
  */
-export default class EventPayload<EventData = Record<string, any>> {
+export default class EventPayload<
+	EventData extends Record<string, any> = Record<string, any>
+> {
 	/**
 	 * Event id.
 	 *
@@ -74,11 +78,11 @@ export default class EventPayload<EventData = Record<string, any>> {
 	 *
 	 * @returns {string}
 	 * @public
-	 * @since 1.0.0
+	 * @since 1.2.0
 	 * @memberof EventPayload
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	public generateId(): string {
-		return (Date.now() * Math.random()).toString() + Math.random().toString(36);
+		return uuidv4();
 	}
 }